♻️  Add "node" feature to have lighter build
When hyle is used as dependency to have model / tools
