+++ src/tools/mod.rs
@@ -1,5 +1,11 @@
 //! Various tools for e.g. profiling and observability.
 
-pub mod contract_runner;
+#[cfg(feature = "node")]
 pub mod mock_workflow;
+
+#[cfg(feature = "tx_builder")]
+pub mod contract_runner;
+#[cfg(feature = "tx_builder")]
 pub mod transactions_builder;
+
+pub mod rest_api_client;
